Fix unsigned byte handling in radix utilities

+++ core/src/commonMain/kotlin/org/gnit/lucenekmp/util/StringSorter.kt
@@ -40,9 +40,9 @@ abstract class StringSorter protected constructor(private val cmp: Comparator<By
             this@StringSorter.swap(i, j)
         }
 
-        override fun byteAt(i: Int, k: Int): Byte {
+        override fun byteAt(i: Int, k: Int): Int {
             get(scratch1, scratchBytes1, i)
-            return cmp.byteAt(scratchBytes1, k).toByte()
+            return cmp.byteAt(scratchBytes1, k)
         }
 
         override fun getFallbackSorter(k: Int): Sorter {
